move trait Scalar
